Move prompt sanitization helpers to text utils
